requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.11.16
    # via -r requirements.in
aiosignal==1.3.2
    # via aiohttp
attrs==25.3.0
    # via aiohttp
av==14.2.0
    # via -r requirements.in
certifi==2025.1.31
    # via requests
cffi==1.17.1
    # via
    #   cryptography
    #   pynacl
    #   soundfile
chardet==5.2.0
    # via -r requirements.in
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via typer
colorama==0.4.6
    # via
    #   click
    #   tqdm
comfyui-frontend-package==1.11.8
    # via -r requirements.in
cryptography==44.0.2
    # via pyjwt
deprecated==1.2.18
    # via pygithub
einops==0.8.1
    # via
    #   -r requirements.in
    #   spandrel
filelock==3.18.0
    # via
    #   huggingface-hub
    #   torch
    #   transformers
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2025.3.2
    # via
    #   huggingface-hub
    #   torch
gitdb==4.0.12
    # via gitpython
gitpython==3.1.44
    # via -r requirements.in
huggingface-hub~=0.29.3
    # via
    #   -r requirements.in
    #   tokenizers
    #   transformers
idna==3.10
    # via
    #   requests
    #   yarl
jinja2==3.1.6
    # via torch
kornia==0.8.0
    # via -r requirements.in
kornia-rs==0.1.8
    # via kornia
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
matrix-client==0.4.0
    # via -r requirements.in
mdurl==0.1.2
    # via markdown-it-py
mpmath==1.3.0
    # via sympy
multidict==6.3.2
    # via
    #   aiohttp
    #   yarl
networkx==3.4.2
    # via torch
numpy==2.2.4
    # via
    #   -r requirements.in
    #   scipy
    #   soundfile
    #   spandrel
    #   torchsde
    #   torchvision
    #   transformers
packaging==24.2
    # via
    #   huggingface-hub
    #   kornia
    #   transformers
pillow==11.1.0
    # via
    #   -r requirements.in
    #   torchvision
propcache==0.3.1
    # via
    #   aiohttp
    #   yarl
psutil==7.0.0
    # via -r requirements.in
pycparser==2.22
    # via cffi
pygithub==2.6.1
    # via -r requirements.in
pygments==2.19.1
    # via rich
pyjwt[crypto]==2.10.1
    # via pygithub
pynacl==1.5.0
    # via pygithub
pyyaml==6.0.2
    # via
    #   -r requirements.in
    #   huggingface-hub
    #   transformers
regex==2024.11.6
    # via transformers
requests==2.32.3
    # via
    #   huggingface-hub
    #   matrix-client
    #   pygithub
    #   transformers
rich==14.0.0
    # via
    #   -r requirements.in
    #   typer
safetensors==0.5.3
    # via
    #   -r requirements.in
    #   spandrel
    #   transformers
scipy==1.15.2
    # via
    #   -r requirements.in
    #   torchsde
sentencepiece==0.2.0
    # via -r requirements.in
shellingham==1.5.4
    # via typer
smmap==5.0.2
    # via gitdb
soundfile==0.13.1
    # via -r requirements.in
spandrel==0.4.1
    # via -r requirements.in
sympy==1.13.1
    # via torch
tokenizers==0.21.1
    # via
    #   -r requirements.in
    #   transformers
toml==0.10.2
    # via -r requirements.in
torch==2.6.0
    # via
    #   -r requirements.in
    #   kornia
    #   spandrel
    #   torchaudio
    #   torchsde
    #   torchvision
torchaudio==2.6.0
    # via -r requirements.in
torchsde==0.2.6
    # via -r requirements.in
torchvision==0.21.0
    # via
    #   -r requirements.in
    #   spandrel
tqdm==4.67.1
    # via
    #   -r requirements.in
    #   huggingface-hub
    #   transformers
trampoline==0.1.2
    # via torchsde
transformers==4.50.3
    # via -r requirements.in
typer==0.15.2
    # via -r requirements.in
typing-extensions==4.13.1
    # via
    #   -r requirements.in
    #   huggingface-hub
    #   pygithub
    #   spandrel
    #   torch
    #   typer
urllib3~=2.3.0
    # via
    #   matrix-client
    #   pygithub
    #   requests
uv==0.6.12
    # via -r requirements.in
wrapt==1.17.2
    # via deprecated
yarl==1.18.3
    # via
    #   -r requirements.in
    #   aiohttp

mysql-connector-python~=9.2.0
pip~=25.0.1
build~=1.2.2.post1
wheel~=0.45.1
xdg~=6.0.0
typing_extensions~=4.13.0
setuptools~=75.8.0